for qiye

+++ codesearch/index/mmap_windows.go
@@ -23,7 +23,7 @@ func mmapFile(f *os.File) mmapData {
 	if size == 0 {
 		return mmapData{f, nil, nil}
 	}
-	h, err := syscall.CreateFileMapping(syscall.Handle(f.Fd()), nil, syscall.PAGE_READONLY, uint32(size>>32), uint32(size), nil)
+	h, err := syscall.CreateFileMapping(syscall.Handle(f.Fd()), nil, syscall.PAGE_READONLY, uint64(size>>32), uint64(size), nil)
 	if err != nil {
 		log.Fatalf("CreateFileMapping %s: %v", f.Name(), err)
 	}
