+++ codesearch/index/read_test.go
@@ -17,8 +17,8 @@ var postFiles = map[string]string{
 	"file3": "Google Web Search",
 }
 
-func tri(x, y, z byte) uint32 {
-	return uint32(x)<<16 | uint32(y)<<8 | uint32(z)
+func tri(x, y, z byte) uint64 {
+	return uint64(x)<<16 | uint64(y)<<8 | uint64(z)
 }
 
 func TestTrivialPosting(t *testing.T) {
@@ -27,27 +27,27 @@ func TestTrivialPosting(t *testing.T) {
 	out := f.Name()
 	buildIndex(out, nil, postFiles)
 	ix := Open(out)
-	if l := ix.PostingList(tri('S', 'e', 'a')); !equalList(l, []uint32{1, 3}) {
+	if l := ix.PostingList(tri('S', 'e', 'a')); !equalList(l, []uint64{1, 3}) {
 		t.Errorf("PostingList(Sea) = %v, want [1 3]", l)
 	}
-	if l := ix.PostingList(tri('G', 'o', 'o')); !equalList(l, []uint32{1, 2, 3}) {
+	if l := ix.PostingList(tri('G', 'o', 'o')); !equalList(l, []uint64{1, 2, 3}) {
 		t.Errorf("PostingList(Goo) = %v, want [1 2 3]", l)
 	}
-	if l := ix.PostingAnd(ix.PostingList(tri('S', 'e', 'a')), tri('G', 'o', 'o')); !equalList(l, []uint32{1, 3}) {
+	if l := ix.PostingAnd(ix.PostingList(tri('S', 'e', 'a')), tri('G', 'o', 'o')); !equalList(l, []uint64{1, 3}) {
 		t.Errorf("PostingList(Sea&Goo) = %v, want [1 3]", l)
 	}
-	if l := ix.PostingAnd(ix.PostingList(tri('G', 'o', 'o')), tri('S', 'e', 'a')); !equalList(l, []uint32{1, 3}) {
+	if l := ix.PostingAnd(ix.PostingList(tri('G', 'o', 'o')), tri('S', 'e', 'a')); !equalList(l, []uint64{1, 3}) {
 		t.Errorf("PostingList(Goo&Sea) = %v, want [1 3]", l)
 	}
-	if l := ix.PostingOr(ix.PostingList(tri('S', 'e', 'a')), tri('G', 'o', 'o')); !equalList(l, []uint32{1, 2, 3}) {
+	if l := ix.PostingOr(ix.PostingList(tri('S', 'e', 'a')), tri('G', 'o', 'o')); !equalList(l, []uint64{1, 2, 3}) {
 		t.Errorf("PostingList(Sea|Goo) = %v, want [1 2 3]", l)
 	}
-	if l := ix.PostingOr(ix.PostingList(tri('G', 'o', 'o')), tri('S', 'e', 'a')); !equalList(l, []uint32{1, 2, 3}) {
+	if l := ix.PostingOr(ix.PostingList(tri('G', 'o', 'o')), tri('S', 'e', 'a')); !equalList(l, []uint64{1, 2, 3}) {
 		t.Errorf("PostingList(Goo|Sea) = %v, want [1 2 3]", l)
 	}
 }
 
-func equalList(x, y []uint32) bool {
+func equalList(x, y []uint64) bool {
 	if len(x) != len(y) {
 		return false
 	}
